Continue to implement environment.
- Implement ability to edit it via ingame gui.
- Split biomes by Wiki types in admin biome selection gui.
- Display biomes available for user in his dimension.
- Add a lot of missing locale strings.
- Change biome cost from int to double. (not supported in editing GUI).

+++ src/main/java/world/bentobox/biomes/database/objects/BiomesObject.java
@@ -161,7 +161,7 @@ public class BiomesObject implements DataObject, Comparable<BiomesObject>
     /**
      * @return the cost of changing biome
      */
-    public int getRequiredCost()
+    public double getRequiredCost()
     {
         return this.requiredCost;
     }
@@ -170,7 +170,7 @@ public class BiomesObject implements DataObject, Comparable<BiomesObject>
     /**
      * @param requiredCost the reqMoney to set
      */
-    public void setRequiredCost(int requiredCost)
+    public void setRequiredCost(double requiredCost)
     {
         this.requiredCost = requiredCost;
     }
@@ -383,7 +383,7 @@ public class BiomesObject implements DataObject, Comparable<BiomesObject>
 
     @ConfigComment("Cost of changing biome.")
     @Expose
-    private int requiredCost;
+    private double requiredCost;
 
     @ConfigComment("Set of String permission that is required for this biome to be activated.")
     @Expose
